Remove TPF #529
