+++ server/src/handlers/mod.rs
@@ -9,6 +9,5 @@ pub mod download;
 pub mod resource;
 pub mod search;
 pub mod single_page_app;
-pub mod tpf;
 pub mod upload;
 pub mod web_sockets;
